requirements.txt
# Project Synth - Requirements
# Updated: November 14, 2025

# ===== Phase 0: Infrastructure (INSTALLED) =====
requests>=2.32.0
python-dotenv>=1.0.0

# ===== Phase 1: Senses - Detection System =====
# Clipboard monitoring (macOS)
pyperclip>=1.8.2
pyobjc-framework-Cocoa>=10.0  # For NSPasteboard

# Screen capture
mss>=9.0.1  # Fast cross-platform screenshots
Pillow>=10.0.0  # Image processing and compression

# ===== Phase 3: Hands - Automation =====
rumps>=0.4.0  # macOS menu bar app
pyobjc-core>=10.0  # macOS automation
pyobjc-framework-AppleScriptKit>=10.0  # AppleScript execution

# ===== Phase 4: Optimization =====
# Fast pre-filtering (to be added)
# sqlite3 (built-in to Python)

# ===== Development =====
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# ===== Optional =====
# py2app>=0.28.0  # Phase 6: Packaging for macOS
